feat(v1.0.0+22): v22 - 球員出席記錄、處罰轉盤、教練訓練頁面、管理員面板、訂閱系統等大量新功能

+++ lib/models/match.dart
@@ -38,7 +38,7 @@ class Match {
     league: json['league'] ?? '',
     homeAway: json['homeAway'] ?? '',
     opponent: json['opponent'] ?? '',
-    jerseyColor: json['jerseyColor'] != null ? Color(json['jerseyColor']) : null,
+    jerseyColor: json['jerseyColor'] != null ? Color(json['jerseyColor'] as int) : null,
   );
 
   Match copyWith({
